Move ports.py
