+++ cli/dstack/_internal/providers/docker/main.py
@@ -4,10 +4,10 @@ from typing import Any, Dict, List, Optional
 from rich_argparse import RichHelpFormatter
 
 import dstack.api.hub as hub
+from dstack._internal.configurators.ports import filter_reserved_ports
 from dstack._internal.core.app import AppSpec
 from dstack._internal.core.job import JobSpec
 from dstack._internal.providers import Provider
-from dstack._internal.providers.ports import filter_reserved_ports
 
 
 class DockerProvider(Provider):
